rearranged in game strategy check
moved the tactic check out of the choose tactic function and into the contestant choose strategy wrapper function for consistency and to simplify the choose tactic function.

also renamed check to tactic_confirmed in the contenstant choose strategy wrapper for clarity, and changed the while loop condition for this from == "n" to != "y" for tighter control.